+++ src/Constants/CompanyTypes.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * This file contains definitions of the available company types.
+ *
+ * Copyright (C) 2022 - today Unzer E-Com GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License")
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @link     https://docs.unzer.com/
+ *
+ * @package  UnzerSDK\Constants
+ */
+
+namespace UnzerSDK\Constants;
+
+class CompanyTypes
+{
+    public const AUTHORITY = 'authority';
+    public const ASSOCIATION = 'association';
+    public const SOLE = 'sole';
+    public const COMPANY = 'company';
+    public const OTHER = 'other';
+}