+++ src/Resources/PaymentTypes/InvoiceSecured.php
@@ -18,8 +18,6 @@
  *
  * @link  https://docs.unzer.com/
  *
- * @author  Simon Gabriel <development@unzer.com>
- *
  * @package  UnzerSDK\PaymentTypes
  */
 namespace UnzerSDK\Resources\PaymentTypes;
@@ -27,6 +25,9 @@ namespace UnzerSDK\Resources\PaymentTypes;
 use UnzerSDK\Traits\CanDirectChargeWithCustomer;
 use UnzerSDK\Traits\IsInvoiceType;
 
+/**
+ * @deprecated since 1.2.0.0 Please use PaylaterInvoice instead.
+ */
 class InvoiceSecured extends BasePaymentType
 {
     use CanDirectChargeWithCustomer;