+++ test/integration/PaymentTypes/BancontactTest.php
@@ -20,11 +20,7 @@
  *
  * @link  https://docs.unzer.com/
  *
- * @author  David Owusu <development@unzer.com>
- * @author  Florian Evertz <development@unzer.com>
- * @author  Simon Gabriel <development@unzer.com>
- *
- * @package  UnzerSDK\test\integration\PaymentTypes
+ @package  UnzerSDK\test\integration\PaymentTypes
  */
 namespace UnzerSDK\test\integration\PaymentTypes;