+++ test/integration/PaymentTypes/PISTest.php
@@ -21,8 +21,6 @@
  *
  * @link  https://docs.unzer.com/
  *
- * @author  Simon Gabriel <development@unzer.com>
- *
  * @package  UnzerSDK\tests/integration/payment_types
  */
 namespace UnzerSDK\test\integration\PaymentTypes;