+++ test/unit/UnzerTest.php
@@ -20,8 +20,6 @@
  *
  * @link  https://docs.unzer.com/
  *
- * @author  Simon Gabriel <development@unzer.com>
- *
  * @package  UnzerSDK\test\unit
  */
 namespace UnzerSDK\test\unit;
@@ -82,7 +80,9 @@ class UnzerTest extends BasePaymentTest
     {
         $unzer = new Unzer('s-priv-1234');
         $unzer->setLocale('myLocale');
+        $unzer->setClientIp('myIpAddress');
         $this->assertEquals('myLocale', $unzer->getLocale());
+        $this->assertEquals('myIpAddress', $unzer->getClientIp());
 
         try {
             $unzer->setKey('this is not a valid key');